refactor: further organize
